rename header and footer test files
