+++ src/integration/pages/idxPage/ukraine_in_russian/amp.test.js
@@ -4,8 +4,8 @@
  */
 
 import runAmpTests from '../ampTests';
-import runHeaderTests from './header.test';
-import runFooterTests from './footer.test';
+import runHeaderTests from './header';
+import runFooterTests from './footer';
 
 describe('AMP ukrainian/ukraine_in_russian IDX page', () => {
   runAmpTests();
