+++ src/integration/pages/idxPage/ukraine_in_russian/canonical.test.js
@@ -4,8 +4,8 @@
  */
 
 import runCanonicalTests from '../canonicalTests';
-import runHeaderTests from './header.test';
-import runFooterTests from './footer.test';
+import runHeaderTests from './header';
+import runFooterTests from './footer';
 
 describe('Canonical ukrainian/ukraine_in_russian IDX page', () => {
   runCanonicalTests();
