+++ src/integration/pages/idxPage/ukraine_in_russian/footer.js
